requirements.txt
discord.py==2.2.2
pytz==2022.7
spotipy==2.22.1
tabulate==0.9.0
yt_dlp==2023.2.17
